nn_menu/ingredients/mcunet/requirements.txt
pandas==1.3.1
onnxruntime==1.8.1
numpy==1.21.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
Pillow==8.3.1
tensorflow==2.5.0